+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -304,8 +304,11 @@ public class Settings extends Config {
 
 	public static ItemData placeholderItem = new ItemData(Material.PAPER);
 
+	public static boolean enableClosingOfPlayerShops = true;
+	public static boolean forceOpenPlayerShops = true;
 	public static ItemData shopOpenItem = new ItemData(Material.GREEN_BANNER);
 	public static ItemData shopClosedItem = new ItemData(Material.RED_BANNER);
+
 	public static ItemData nameItem = new ItemData(Material.NAME_TAG);
 
 	public static boolean enableAllEquipmentEditorSlots = false;